Adjust types for 32-bit
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@317 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/_mingw.h
@@ -223,7 +223,7 @@ extern "C" {
   typedef unsigned __int64 size_t;
 #endif
 #else
-  typedef unsigned long size_t;
+  typedef unsigned int size_t;
 #endif
 #endif
 
@@ -237,7 +237,7 @@ extern "C" {
   typedef __int64 ssize_t;
 #endif
 #else
-  typedef long ssize_t;
+  typedef int ssize_t;
 #endif
 #endif
 
@@ -251,7 +251,7 @@ extern "C" {
   typedef __int64 intptr_t;
 #endif
 #else
-  typedef long intptr_t;
+  typedef int intptr_t;
 #endif
 #endif
 
@@ -265,7 +265,7 @@ extern "C" {
   typedef unsigned __int64 uintptr_t;
 #endif
 #else
-  typedef unsigned long uintptr_t;
+  typedef unsigned int uintptr_t;
 #endif
 #endif
 
@@ -280,7 +280,7 @@ extern "C" {
   typedef __int64 ptrdiff_t;
 #endif
 #else
-  typedef long ptrdiff_t;
+  typedef int ptrdiff_t;
 #endif
 #endif
 #endif
